backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]
sqlalchemy
mysql-connector-python